+++ how-to/automation-testing/register-with-store-js/tests/index.spec.js
@@ -15,7 +15,7 @@ describe('Register with Store', () => {
 	it('The runtime version should be set', async () => {
 		const fin = await OpenFinProxy.fin();
 		const version = await fin.System.getVersion();
-		expect(version).to.equal('30.110.74.7');
+		expect(version).to.equal('30.110.74.8');
 	});
 
 	it('Can open the store window', async () => {