+++ how-to/customize-workspace/client/src/framework/apps.ts
@@ -5,6 +5,7 @@ import type {
 	AppEndpointOptions,
 	AppFilterOptions,
 	AppProviderOptions,
+	AppsForIntent,
 	PlatformApp
 } from "./shapes/app-shapes";
 import type { EndpointProvider } from "./shapes/endpoint-shapes";
@@ -364,9 +365,7 @@ export async function getIntent(
 	return null;
 }
 
-export async function getIntentsByContext(
-	contextType: string
-): Promise<{ intent: { name: string; displayName: string }; apps: PlatformApp[] }[]> {
+export async function getIntentsByContext(contextType: string): Promise<AppsForIntent[]> {
 	const apps = await getApps();
 	let intents: {
 		[key: string]: {