+++ how-to/customize-workspace/client/src/framework/launch.ts
@@ -4,64 +4,41 @@ import { launchConnectedApp } from "./connections";
 import * as endpointProvider from "./endpoint";
 import { createLogger } from "./logger-provider";
 import { manifestTypes } from "./manifest-types";
+import {
+	bringViewToFront,
+	bringWindowToFront,
+	doesViewExist,
+	doesWindowExist,
+	findViewNames
+} from "./platform/browser";
 import { getSettings } from "./settings";
-import type { PlatformApp } from "./shapes/app-shapes";
+import type { PlatformApp, PlatformAppIdentifier } from "./shapes/app-shapes";
 import { randomUUID } from "./uuid";
 
 const logger = createLogger("Launch");
 
-async function getViewIdentities(name: string, uuid: string) {
+async function getViewIdentities(
+	name: string,
+	uuid: string,
+	appId: string
+): Promise<PlatformAppIdentifier[]> {
 	const identity = { uuid, name };
 	const win = fin.Window.wrapSync(identity);
 	const views = await win.getCurrentViews();
-	const viewIdentities = views.map((view) => view.identity);
+	const viewIdentities = views.map((view) => ({ ...view.identity, appId }));
 	await win.setAsForeground();
 	return viewIdentities;
 }
 
-async function doesViewExist(name: string, uuid: string) {
-	const view = fin.View.wrapSync({ name, uuid });
-	let exists = false;
-	try {
-		await view.getInfo();
-		const viewHost = await view.getCurrentWindow();
-		await viewHost.bringToFront();
-		exists = true;
-	} catch {
-		exists = false;
+function decoupleApp(app: PlatformApp) {
+	if (app === undefined || app === null) {
+		logger.warn("No app was passed to decoupleApp");
+		return app;
 	}
-	return exists;
+	return JSON.parse(JSON.stringify(app)) as PlatformApp;
 }
 
-async function doesWindowExist(name: string, uuid: string) {
-	const win = fin.Window.wrapSync({ name, uuid });
-	let exists = false;
-	try {
-		await win.getInfo();
-		exists = true;
-		if (await win.isShowing()) {
-			await win.bringToFront();
-		}
-	} catch {
-		exists = false;
-	}
-	return exists;
-}
-
-function findViewNames(layout) {
-	const collectedNames: string[] = [];
-
-	JSON.stringify(layout, (_, nestedValue) => {
-		// check to ensure that we have a name field and that we also have a url field in this object (in case name was added to a random part of the layout)
-		if (nestedValue?.name?.length && nestedValue.url !== undefined) {
-			collectedNames.push(nestedValue.name as string);
-		}
-		return nestedValue as unknown;
-	});
-
-	return collectedNames;
-}
-export async function launchWindow(windowApp: PlatformApp): Promise<OpenFin.Identity> {
+async function launchWindow(windowApp: PlatformApp): Promise<PlatformAppIdentifier> {
 	if (windowApp === undefined || windowApp === null) {
 		logger.warn("No app was passed to launchWindow");
 		return null;
@@ -87,31 +64,44 @@ export async function launchWindow(windowApp: PlatformApp): Promise<OpenFin.Iden
 		manifest = windowApp.manifest as unknown as OpenFin.WindowOptions;
 	}
 
-	const name = manifest.name;
+	let name = manifest.name;
+	let wasNameSpecified = name !== undefined;
+
+	if (!wasNameSpecified && windowApp?.customConfig?.instanceMode === "single") {
+		logger.info(
+			`A unique name was not provided in the manifest of this window but the custom config indicates that this app is supposed to have a single instance so we are using the appId: ${windowApp.appId} as the unique name.`
+		);
+		name = windowApp.appId;
+		// assign the new name to the manifest in case the view doesn't exist yet (as it is dynamically allocated and not part of the app's manifest)
+		manifest.name = name;
+		wasNameSpecified = true;
+	} else if (!wasNameSpecified) {
+		name = `${windowApp.appId}/${randomUUID()}`;
+		manifest.name = name;
+	}
+
 	let identity = { uuid: fin.me.identity.uuid, name };
-	const wasNameSpecified = name !== undefined;
+
 	let windowExists = false;
 
 	if (wasNameSpecified) {
-		windowExists = await doesWindowExist(identity.name, identity.uuid);
-	} else {
-		manifest.name = `classic-window-${randomUUID()}`;
-		identity.name = manifest.name;
+		windowExists = await doesWindowExist(identity.name, identity.uuid, true);
 	}
 
 	if (!windowExists) {
 		try {
 			const createdWindow = await fin.Window.create(manifest);
 			identity = createdWindow.identity;
+			await bringWindowToFront({ window: createdWindow });
 		} catch (err) {
 			logger.error("Error launching window", err);
 			return null;
 		}
 	}
-	return identity;
+	return { ...identity, appId: windowApp.appId };
 }
 
-export async function launchView(viewApp: PlatformApp): Promise<OpenFin.Identity> {
+async function launchView(viewApp: PlatformApp): Promise<PlatformAppIdentifier> {
 	if (viewApp === undefined || viewApp === null) {
 		logger.warn("No app was passed to launchView");
 		return null;
@@ -137,13 +127,29 @@ export async function launchView(viewApp: PlatformApp): Promise<OpenFin.Identity
 		manifest = viewApp.manifest as unknown as OpenFin.ViewOptions;
 	}
 
-	const name = manifest.name;
+	let name = manifest.name;
+
+	let wasNameSpecified = name !== undefined;
+
+	if (!wasNameSpecified && viewApp?.customConfig?.instanceMode === "single") {
+		logger.info(
+			`A unique name was not provided in the manifest of this view but the custom config indicates that this app is supposed to have a single instance so we are using the appId: ${viewApp.appId} as the unique name.`
+		);
+		name = viewApp.appId;
+		// assign the new name to the manifest in case the view doesn't exist yet (as it is dynamically allocated and not part of the app's manifest)
+		manifest.name = name;
+		wasNameSpecified = true;
+	} else if (!wasNameSpecified) {
+		name = `${viewApp.appId}/${randomUUID()}`;
+		manifest.name = name;
+	}
+
 	let identity = { uuid: fin.me.identity.uuid, name };
-	const wasNameSpecified = name !== undefined;
+
 	let viewExists = false;
 
 	if (wasNameSpecified) {
-		viewExists = await doesViewExist(identity.name, identity.uuid);
+		viewExists = await doesViewExist(identity.name, identity.uuid, true);
 	}
 
 	if (!viewExists) {
@@ -156,10 +162,10 @@ export async function launchView(viewApp: PlatformApp): Promise<OpenFin.Identity
 			return null;
 		}
 	}
-	return identity;
+	return { ...identity, appId: viewApp.appId };
 }
 
-export async function launchSnapshot(snapshotApp: PlatformApp): Promise<OpenFin.Identity[]> {
+async function launchSnapshot(snapshotApp: PlatformApp): Promise<PlatformAppIdentifier[]> {
 	if (snapshotApp === undefined || snapshotApp === null) {
 		logger.warn("No app was passed to launchSnapshot");
 		return null;
@@ -178,27 +184,27 @@ export async function launchSnapshot(snapshotApp: PlatformApp): Promise<OpenFin.
 
 	if (Array.isArray(windows)) {
 		const windowsToGather: string[] = [];
-		const viewIds: OpenFin.Identity[] = [];
+		const viewIds: PlatformAppIdentifier[] = [];
 
-		for (let i = 0; i < windows.length; i++) {
-			const getViewIdsForLayout = findViewNames(windows[i].layout);
+		for (const currentWindow of windows) {
+			const getViewIdsForLayout = findViewNames(currentWindow.layout);
 			if (getViewIdsForLayout.length === 0) {
 				const uuid = randomUUID();
-				const name = `internal-generated-window-${uuid}`;
-				windows[i].name = name;
-				windowsToCreate.push(windows[i]);
+				const name = `${snapshotApp.appId}/${uuid}`;
+				currentWindow.name = name;
+				windowsToCreate.push(currentWindow);
 				windowsToGather.push(name);
 			} else {
 				// we have views. Grab the first one to validate existence.
 				const viewId = getViewIdsForLayout[0];
 
 				for (const entry of getViewIdsForLayout) {
-					viewIds.push({ name: entry, uuid: fin.me.identity.uuid });
+					viewIds.push({ name: entry, uuid: fin.me.identity.uuid, appId: snapshotApp.appId });
 				}
 
 				// these views should be readonly and cannot be pulled out of the page or closed
-				if (!(await doesViewExist(viewId, fin.me.identity.uuid))) {
-					windowsToCreate.push(windows[i]);
+				if (!(await doesViewExist(viewId, fin.me.identity.uuid, false))) {
+					windowsToCreate.push(currentWindow);
 				}
 			}
 		}
@@ -214,8 +220,8 @@ export async function launchSnapshot(snapshotApp: PlatformApp): Promise<OpenFin.
 			}
 		}
 
-		for (let w = 0; w < windowsToGather.length; w++) {
-			const windowViewIds = await getViewIdentities(windowsToGather[w], fin.me.identity.uuid);
+		for (const targetWindow of windowsToGather) {
+			const windowViewIds = await getViewIdentities(targetWindow, fin.me.identity.uuid, snapshotApp.appId);
 			viewIds.push(...windowViewIds);
 		}
 
@@ -225,73 +231,209 @@ export async function launchSnapshot(snapshotApp: PlatformApp): Promise<OpenFin.
 	return null;
 }
 
-export async function launch(appEntry: PlatformApp) {
+export async function launch(platformApp: PlatformApp): Promise<PlatformAppIdentifier[]> {
 	try {
-		logger.info("Application launch requested", appEntry);
-		if (appEntry.manifestType === manifestTypes.external.id) {
-			const settings = await getSettings();
-			const appAssetTag = settings?.appProvider?.appAssetTag ?? "appasset";
+		logger.info("Application launch requested", platformApp);
 
-			if (appEntry.tags?.includes(appAssetTag)) {
-				logger.info(
-					"Application requested is a native app with a tag of appasset so it is provided by this workspace platform. Managing request via platform and not Workspace"
-				);
+		if (platformApp === undefined || platformApp === null) {
+			logger.warn("An empty app definition was passed to launch");
+			return [];
+		}
+		const app = decoupleApp(platformApp);
+
+		const platformAppIdentities: PlatformAppIdentifier[] = [];
+		switch (app.manifestType) {
+			case manifestTypes.external.id: {
+				const settings = await getSettings();
+				const appAssetTag = settings?.appProvider?.appAssetTag ?? "appasset";
 				const options: OpenFin.ExternalProcessRequestType = {};
-				options.alias = appEntry.manifest;
-				options.uuid = appEntry.appId;
 
-				await fin.System.launchExternalProcess(options);
-			} else {
-				logger.info("Application requested is a native app. Managing request via platform and not Workspace");
-				const options: OpenFin.ExternalProcessRequestType = {};
-				options.path = appEntry.manifest;
-				options.uuid = appEntry.appId;
-
-				await fin.System.launchExternalProcess(options);
-			}
-		} else if (appEntry.manifestType === manifestTypes.inlineView.id) {
-			await launchView(appEntry);
-		} else if (
-			appEntry.manifestType === manifestTypes.window.id ||
-			appEntry.manifestType === manifestTypes.inlineWindow.id
-		) {
-			await launchWindow(appEntry);
-		} else if (appEntry.manifestType === manifestTypes.inlineExternal.id) {
-			logger.info(
-				"Application requested is a native app defined as inline-external. Managing request via platform and not Workspace."
-			);
-			try {
-				const options = appEntry.manifest as OpenFin.ExternalProcessRequestType;
-				await fin.System.launchExternalProcess(options);
-			} catch (err) {
-				logger.error(`Error trying to launch inline-external with appId: ${appEntry.appId}`, err);
-			}
-		} else if (appEntry.manifestType === manifestTypes.desktopBrowser.id) {
-			await fin.System.openUrlWithBrowser(appEntry.manifest);
-		} else if (appEntry.manifestType === manifestTypes.endpoint.id) {
-			if (endpointProvider.hasEndpoint(appEntry.manifest)) {
-				const launched = await endpointProvider.action(appEntry.manifest, { payload: appEntry });
-				if (!launched) {
-					logger.warn(
-						`App with id: ${appEntry.appId} encountered when launched using endpoint: ${appEntry.manifest}.`
+				if (app.tags?.includes(appAssetTag)) {
+					logger.info(
+						`Application requested is a native app with a tag of ${appAssetTag} so it is provided by this workspace platform. Managing request via platform and not Workspace`
 					);
+					options.alias = app.manifest;
+					options.uuid = app.appId;
+				} else {
+					logger.info(
+						"Application requested is a native app. Managing request via platform and not Workspace"
+					);
+					options.path = app.manifest;
+					options.uuid = app.appId;
+				}
+				const identity = await fin.System.launchExternalProcess(options);
+				platformAppIdentities.push({ ...identity, appId: app.appId });
+				break;
+			}
+			case manifestTypes.inlineExternal.id: {
+				logger.info(
+					"Application requested is a native app defined as inline-external. Managing request via platform and not Workspace."
+				);
+				try {
+					const options = app.manifest as OpenFin.ExternalProcessRequestType;
+					const identity = await fin.System.launchExternalProcess(options);
+					platformAppIdentities.push({ ...identity, appId: app.appId });
+				} catch (err) {
+					logger.error(`Error trying to launch inline-external with appId: ${app.appId}`, err);
+				}
+				break;
+			}
+			case manifestTypes.inlineView.id:
+			case manifestTypes.view.id: {
+				const platformIdentity = await launchView(app);
+				if (platformIdentity !== null) {
+					platformAppIdentities.push(platformIdentity);
+				}
+				break;
+			}
+			case manifestTypes.window.id:
+			case manifestTypes.inlineWindow.id: {
+				const platformIdentity = await launchWindow(app);
+				if (platformIdentity !== null) {
+					platformAppIdentities.push(platformIdentity);
+				}
+				break;
+			}
+			case manifestTypes.snapshot.id: {
+				const identities = await launchSnapshot(app);
+				if (identities !== null && identities !== undefined) {
+					platformAppIdentities.push(...identities);
+				}
+				break;
+			}
+			case manifestTypes.manifest.id: {
+				const manifest = await fin.System.launchManifest(app.manifest);
+				if (manifest?.platform?.uuid !== undefined) {
+					platformAppIdentities.push({
+						uuid: manifest.platform.uuid,
+						name: manifest.platform.uuid,
+						appId: app.appId
+					});
+				} else if (manifest?.startup_app?.uuid !== undefined) {
+					platformAppIdentities.push({
+						uuid: manifest.startup_app.uuid,
+						name: manifest.startup_app.uuid,
+						appId: app.appId
+					});
+				}
+				break;
+			}
+			case manifestTypes.desktopBrowser.id: {
+				await fin.System.openUrlWithBrowser(app.manifest);
+				break;
+			}
+			case manifestTypes.endpoint.id: {
+				if (endpointProvider.hasEndpoint(app.manifest)) {
+					const identity = await endpointProvider.requestResponse<{ payload: PlatformApp }, OpenFin.Identity>(
+						app.manifest,
+						{ payload: app }
+					);
+					if (identity === undefined || identity === null) {
+						logger.warn(
+							`App with id: ${app.appId} encountered when launched using endpoint: ${app.manifest}.`
+						);
+					} else {
+						platformAppIdentities.push({ ...identity, appId: app.appId });
+					}
+				} else {
+					logger.warn(
+						`App with id: ${app.appId} could not be launched as it is of manifestType: ${app.manifestType} and the endpoint: ${app.manifest} is not available.`
+					);
+				}
+				break;
+			}
+			case manifestTypes.connection.id: {
+				logger.info(
+					"An app defined by a connection (connected app) has been selected. Passing selection to connection"
+				);
+				const identity = await launchConnectedApp(app);
+				if (identity !== undefined && identity !== null) {
+					platformAppIdentities.push({ ...identity, appId: app.appId });
+				}
+				break;
+			}
+			default: {
+				logger.error("We do not support the manifest type so this app cannot be launched.", app);
+			}
+		}
+
+		logger.info("Finished application launch request");
+		return platformAppIdentities;
+	} catch (err) {
+		logger.error("Failed during application launch request", err);
+	}
+}
+
+export async function bringToFront(appEntry: PlatformApp, platformAppIdentifiers: PlatformAppIdentifier[]) {
+	switch (appEntry.manifestType) {
+		case manifestTypes.external.id:
+		case manifestTypes.inlineExternal.id:
+		case manifestTypes.manifest.id:
+		case manifestTypes.desktopBrowser.id:
+		case manifestTypes.endpoint.id:
+		case manifestTypes.connection.id: {
+			logger.info(`Bringing apps of type: ${appEntry.manifestType} to front is not supported.`);
+			break;
+		}
+		case manifestTypes.inlineView.id:
+		case manifestTypes.view.id: {
+			if (Array.isArray(platformAppIdentifiers) && platformAppIdentifiers.length === 1) {
+				await bringViewToFront({ identity: platformAppIdentifiers[0] });
+			} else {
+				logger.warn(
+					`A request to bring a view app to front was received but we didn't receive exactly one identity for app: ${appEntry.appId}.`
+				);
+			}
+			break;
+		}
+		case manifestTypes.window.id:
+		case manifestTypes.inlineWindow.id: {
+			if (Array.isArray(platformAppIdentifiers) && platformAppIdentifiers.length === 1) {
+				await bringWindowToFront({ identity: platformAppIdentifiers[0] });
+			} else {
+				logger.warn(
+					`A request to bring a window app to front was received but we didn't receive exactly one identity for app: ${appEntry.appId}.`
+				);
+			}
+			break;
+		}
+		case manifestTypes.unregisteredApp.id: {
+			if (Array.isArray(platformAppIdentifiers) && platformAppIdentifiers.length === 1) {
+				// isView - this should be the majority of cases for an unregistered app
+				let isView = true;
+				try {
+					const view = fin.View.wrapSync(platformAppIdentifiers[0]);
+					const viewInfo = await view.getInfo();
+					logger.info("View Info", viewInfo);
+				} catch {
+					isView = false;
+				}
+				if (isView) {
+					await bringViewToFront({ identity: platformAppIdentifiers[0] });
+				} else {
+					await bringWindowToFront({ identity: platformAppIdentifiers[0] });
 				}
 			} else {
 				logger.warn(
-					`App with id: ${appEntry.appId} could not be launched as it is of manifestType: ${appEntry.manifestType} and the endpoint: ${appEntry.manifest} is not available.`
+					`A request to bring a window app to front was received but we didn't receive exactly one identity for app: ${appEntry.appId}.`
 				);
 			}
-		} else if (appEntry.manifestType === manifestTypes.connection.id) {
-			logger.info(
-				"An app defined by a connection (connected app) has been selected. Passing selection to connection"
-			);
-			await launchConnectedApp(appEntry);
-		} else {
-			const platform = getCurrentSync();
-			await platform.launchApp({ app: appEntry });
+			break;
 		}
-		logger.info("Finished application launch request");
-	} catch (err) {
-		logger.error("Failed during application launch request", err);
+		case manifestTypes.snapshot.id: {
+			if (Array.isArray(platformAppIdentifiers) && platformAppIdentifiers.length > 0) {
+				for (const identity of platformAppIdentifiers) {
+					await bringViewToFront({ identity });
+				}
+			} else {
+				logger.warn(
+					`A request to bring a snapshot app to front was received but we didn't receive at least one view for app: ${appEntry.appId}.`
+				);
+			}
+			break;
+		}
+		default: {
+			logger.error("We do not support the manifest type so this app cannot be brought to front.", appEntry);
+		}
 	}
 }